+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 
 **Changed**
 
+- #70 Index and Metadata optimizations for AR Listing View
 
 **Removed**
 
@@ -16,6 +17,7 @@ Changelog
 **Fixed**
 
 - #71 Slow Reference Widgets in 'Add Case' View
+- #69 Add proper license to repository
 
 **Security**